pyproject.toml
[project]
name = "gen-ai-project"
version = "0.1.0"
description = "A Gen AI project"
requires-python = ">=3.9"
dependencies = [
    "google-generativeai>=0.3.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.28.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/gen_ai_project"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
